+++ .gitignore
@@ -24,6 +24,7 @@ yarn-error.log*
 
 /src/to-be-sorted/
 AppTest.jsx
+VoteMainTest.jsx
 *secret*
 
 # https://stackoverflow.com/questions/67626696/radio-buttons-with-react-hook-form
