+++ package.json
@@ -19,7 +19,9 @@
     "react-hook-form": "^7.45.2",
     "react-responsive-carousel": "^3.2.23",
     "react-scripts": "5.0.1",
+    "react-select": "^5.7.4",
     "react-share": "^4.4.1",
+    "react-tabs": "^6.0.2",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
