+++ src/components/VoteMain.jsx
@@ -1,5 +1,6 @@
 import { useEffect, useState } from "react";
 import "./css/Poll.css";
+import { VotingSection } from "./VotingSection";
 
 function VoteMain(props) {
     
@@ -8,13 +9,6 @@ function VoteMain(props) {
     const [selectedParticipatingPollId, setSelectedParticipatingPollId] = useState(-1)
     const [participatingPollsLoading, setParticipitingPollsLoading] = useState(true)
     const [totalParticipatingPolls, setTotalParticipatingPolls] = useState(0)
-
-    const [votingSection, setVotingSection] = useState([])
-    const [pollQuestion, setPollQuestion] = useState();
-    const [pollOptions, setPollOptions] = useState([]);
-    const [pollCount, setPollCount] = useState(0)
-    const [loading, setLoading] = useState(true);
-    const [selectedOption, setSelectedOption] = useState(-1);
 
     // GET API Calls
     const loadParticipatingPolls = async () => {
@@ -39,104 +33,11 @@ function VoteMain(props) {
             setParticipitingPollsLoading(false)
           })
     }
-
-    const loadPolls = async () => {
-      await fetch("http://localhost:3003/votesection", {
-        method: "GET",
-        headers: {
-          "Accept": "application/json",
-          "Content-Type": "application/json",
-          "Token": sessionStorage.getItem("user"),
-        },
-      })
-        .then((res) => res.json())
-        .then((res) => {
-          // console.log("vote section: ", res["data"])
-          setPollCount(res["data"].length)
-          
-          setVotingSection(res["data"])
-          setLoading(false);
-
-        });
-    };
-    const handleCancelClick = (e) => {
-      setSelectedOption(-1);
-    };
-    const handleSubmitClick = async (e) => {
-      // console.log()
-      // const selectedVotingSection = votingSection.slice()
-      let pollId = parseInt(e.currentTarget.attributes.poll_id.value)
-      let voteId = parseInt(e.currentTarget.attributes.vote_id.value)
-      let selectedVoteId = votingSection.filter((vs) => vs.poll_id === parseInt(e.currentTarget.attributes.poll_id.value))[0].data.filter((p) => p.vote_id === parseInt(e.currentTarget.attributes.vote_id.value))[0].selected_vote_detail_id
-      
-      console.log(parseInt(e.currentTarget.attributes.poll_id.value), 
-                  parseInt(e.currentTarget.attributes.vote_id.value), 
-                  votingSection.filter((vs) => vs.poll_id === parseInt(e.currentTarget.attributes.poll_id.value))[0].data.filter((p) => p.vote_id === parseInt(e.currentTarget.attributes.vote_id.value))[0].selected_vote_detail_id)
-      
-      await fetch("http://localhost:3003/savevote",
-      {
-        method: "POST",
-        headers: {
-          "Accept": "application/json",
-          "Content-Type": "application/json",
-          "Token": sessionStorage.getItem("user")
-        },
-        body: JSON.stringify(
-          {
-            poll_id : pollId,
-            vote_id : voteId,
-            selected_vote_id : selectedVoteId
-          }
-        )
-      }
-    ).then((res) => res.json()).then(res => {
-      console.log(res)
-      console.log(res["data"])
-      
-      // setMyGroups(res["data"])
-      // setLoadingMyGroups(false);
-      // setTotalMyGroups(res["data"].length)
-
-      // setGroupName('')
-    });
-      // .filter((p) => p.vote_id === parseInt(e.currentTarget.attributes.vote_id.value))
-      // console.log(selectedVotingSection.filter((vs) => vs.poll_id = parseInt(e.currentTarget.attributes.poll_id.value)))
-      // console.log(e.target.id)
-      // console.log("handleSubmitMe: poll_id / vote_id: ", e.target.attributes.poll_id.value + " / " + e.target.attributes.vote_id.value)
-      // console.log("submit:", votingResults.filter(vr => vr.poll_id ===  parseInt(e.target.attributes.poll_id.value) && vr.vote_id === parseInt(e.target.attributes.vote_id.value))[0].vote_detail_id)
-    };
-
-    // const handleSubmitForm = (e) => {
-    //   e.preventDefault()
-    //   console.clear()
-    //   console.log("handle Submit Click Form Start")
-    //   console.log("e: ", e)
-    //   console.log("e.target: ", e.target)
-    //   console.log("e.target: ", e.targetform)
-    //   // console.log(inputs)
-    //   console.log("handle Submit Click Form End")
-    // }
-
-    const handleOptionClick = (e) => {
-      const updatedVotingSection = votingSection.slice()
-      // console.log("votingSection before :", votingSection)
-      updatedVotingSection.map((vs) => {vs.poll_id === parseInt(e.target.attributes.poll_id.value) ?
-                                            vs.data.map((d) => {d.selected_vote_detail_id = d.vote_id === parseInt(e.target.attributes.vote_id.value) ?
-                                                parseInt(e.target.id) : d.selected_vote_detail_id}):
-                                        void 0})
-
-      setVotingSection(updatedVotingSection)
-      // console.log("votingSection after :", votingSection)
-    };
   
     useEffect(() => {
       setParticipatingPolls([]);
       setParticipitingPollsLoading(true);
       loadParticipatingPolls()
-      setPollOptions([]);
-      setPollQuestion();
-      setLoading(true);
-      loadPolls();
     }, [props]);
   
     const handleNavigate = (e) => {
@@ -198,103 +99,8 @@ function VoteMain(props) {
               {/* Participating Poll End */}
 
               {/* Voting Section Start */}
-              <>                
-                <div className="activePollOptionContainer">
-                  {votingSection.map((vote) => {
-                      return (
-                        (selectedParticipatingPollId === vote.poll_id &&
-                          <div key={vote.poll_id}>
-                            {
-                              vote.data.map((vd) => {
-                                return (
-                                  <div key={vd.vote_title} >
-                                    {/* <form tag = {vd.vote_title} onSubmit={handleSubmitForm}> */}
-                                    <div className="activePollTitle">{vd.vote_title}</div>
-                                    {vd.vote_detail.map((vdd) => {
-                                      // console.log(vote.poll_id, vd.vote_id)
-                                      // console.log(votingResults)
-                                      // console.log(votingResults.filter(vr => vr.poll_id === vote.poll_id))
-                                      
-                                      // console.log(votingResults.filter(vr => vr.poll_id === vote.poll_id && vr.vote_id === vd.vote_id)[0].vote_detail_id === vdd.vote_detail_id ? true : false)
-                                      // console.log(votingResults.filter(vr => vr.poll_id === vote.poll_id && vr.vote_id === vd.vote_id))
-                                      return(
-                                          <div key={vdd.vote_detail_id} className="activePollOptions">
-                                            <div className="form-check">
-                                              <input
-                                                className="form-check-input"
-                                                type="radio"
-                                                name={vd.vote_title}
-                                                value={vdd.option}
-                                                id={vdd.vote_detail_id}
-                                                poll_id = {vote.poll_id}
-                                                vote_id = {vd.vote_id}
-                                                // checked={vdd.vote_detail_id === selectedOption}
-                                                // checked={votingResults.filter(vr => vr.poll_id === vote.poll_id && vr.vote_id === vd.vote_id)[0].vote_detail_id === vdd.vote_detail_id ? true : false}
-                                                checked={vdd.vote_detail_id === vd.selected_vote_detail_id ? true : false}
-                                                onChange={handleOptionClick}
-                                                option = {vdd['option']}
-                                              />
-                                              <label
-                                                className="form-check-label"
-                                                htmlFor="flexRadioDefault1"
-                                              >
-                                                {vdd.option}
-                                              </label>
-                                            </div>
-                                          </div>
-                                      )}
-                                      )
-                                    }
-                                    <div className="activePollActionsContainer">
-                                      <button
-                                        name = {vd.vote_id.toString()}
-                                        id = {vd.vote_id.toString()} 
-                                        poll_id = {vote.poll_id}
-                                        vote_id = {vd.vote_id}
-                                        value = {vd.vote_id}
-                                        className="activePollActionButton"
-                                        onClick={handleSubmitClick}
-                                        type="submit"
-                                      >
-                                        <span className="pollActionText"> Submit </span>
-                                        {/* Submit */}
-                                      </button>
-                                      {/* <button 
-                                        name = {vd.vote_id.toString()}
-                                        id = {vd.vote_id.toString()} 
-                                        poll_id = {vote.poll_id}
-                                        vote_id = {vd.vote_id}
-                                        value = {vd.vote_id}
-                                        className="activePollActionButton"
-                                        onClick={handleSubmitClick}
-                                        type="submit"
-                                      >
-                                        Submit 2
-                                      </button> */}
-                                      
-                                      {/* <button
-                                        className="activePollActionButton"
-                                        onClick={handleCancelClick}
-                                        disabled={pollCount === 0}
-                                      >
-                                        <span className="pollActionText"> Cancel </span>
-                                      </button> */}
-
-                                      {/* <input type="submit" value = "submit form"/> */}
-                                    
-                                    </div>
-                                    
-                                    {/* </form> */}
-                                  </div>
-                                )
-                              })
-                            }
-                        </div>
-                        )
-                      )
-                    })
-                  }
-                </div>
+              <>
+                <VotingSection selectedParticipatingPollId={selectedParticipatingPollId}/>
               </>
               {/* Voting Section End */}
             </>
@@ -305,3 +111,4 @@ function VoteMain(props) {
   }
 
   export {VoteMain}
+
