+++ src/components/css/Poll.css
@@ -424,3 +424,67 @@
     border-top-right-radius: 20px;
     border-bottom-right-radius: 20px;
 }
+.adminFreezeButton {
+    padding-top: 10px;
+    margin-top: 10px;
+    padding-left: 10px;
+    padding-right: 10px;
+    /* width: 150px; */
+}
+.adminRightAnswerDropDown{
+    font-size: small;
+    color: #0c0c0c;
+    /* width: 100%; */
+    height: auto;
+    /* padding-left: 10px; */
+    width: 150px;
+}
+.Tabs {
+    /* width: 80%; */
+    /* height: auto; */
+    min-height: 10px;
+    background: #141616;
+    /* margin: 3.5rem auto 1.5rem; */
+    /* padding: 2rem; */
+    color: #ecece8;
+    /* border-radius: 2rem; */
+    @media (max-width: 769px) {
+     padding: 2rem 0;
+    }
+   }
+
+   ul.nav {
+    /* width: 60%; */
+    margin: 0 auto 0rem;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    border: 1px solid #39A2DB;
+    border-radius: 2rem;
+    padding-left: 0px;
+    @media (max-width: 768px) {
+      width: 90%;
+    }
+  }
+  
+  ul.nav li {
+    width: 50%;
+    padding: 0.25rem;
+    list-style: none;
+    text-align: center;
+    cursor: pointer;
+    transition: all 0.7s;
+    border-bottom-left-radius: 2rem;
+    border-top-left-radius: 2rem;
+  }
+  ul.nav li:nth-child(2) {
+    border-radius: 0;
+    border-bottom-right-radius: 2rem;
+    border-top-right-radius: 2rem;
+  }
+  ul.nav li:hover {
+    background: rgba(152, 230, 217, 0.15);
+  }
+  ul.nav li.active {
+    background: #5c88a2;
+  }
